requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
Pillow==10.1.0
openai==2.3.0
boto3==1.34.0
botocore==1.34.0
reportlab==4.0.7
python-dotenv==1.0.0
opencv-python==4.10.0.84
numpy<2.0.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.8.2
seaborn==0.13.0
webcolors==1.13
scikit-learn==1.3.2
colorama==0.4.6
scikit-learn
boto3 
botocore